+++ render/views/__tests__/AppViewPages/IssueDetailsPage.spec.jsx
@@ -99,6 +99,9 @@ const state = {
   },
   fields: {
     data: {}
+  },
+  session: {
+    statusBar: ''
   }
 };